Make the snake move by itself.
However, checking the endgame condition should happen in the model itself,
instead of in SnakeEvents. Since the model is the one keeping track of the game,
it should know by itself if endgame has been reached. As it stands, if the snake
collapses with the wall due to automove, endgame will not be declared since
checking is in SnakeEvents.