Sketch savepoint for explorable.
I need a clean way to translate the sprites to simulate a moving world. I'm
hoping to rely on the `update` method of a `ForegroundSprite` but I need a way
to introduce some structure in the method call; it's unrealistic to expect that
only `Explorable` will ever call said method.