Make monster shooter tests pass. But it is not without problems!
First off, it seems that the windows are not closing even after endgame
condition (as it should! See behavior when the games are actually played, not
merely unit-tested). However, previous iterations (e.g., Snake), _did_ close
their windows after endgame condition was reached. So now I wonder why it is
not doing that.

Also, printed when `pygame.init()` is called and looks like that, at least for
monster shooter, it is only called _after the window is closed_.