Version bump to v1.0.0.
WHOA VERSION 1.0.0.

Incrementing to 1.0.0 is not meant to advertise that this framework is ready for
primetime. Instead, it is done purely to comply with semantic versioning schema.
This version is not backwards compatible since it removes the ugly
`KeyboardHandlerMapping` in favor of `KeyControls`.

To indicate that this is still very much a work-in-progress, some stray prints
are (un)intentionally left in the code.